Divided analyzing logic into submodules.
