+++ analyze/analyses/mod.rs
@@ -1,0 +1,6 @@
+pub mod diff;
+pub mod dominators;
+pub mod garbage;
+pub mod monos;
+pub mod paths;
+pub mod top;
